cleanup: convert tabs to spaces
This codebase originated on an embedded platform with an established
codebase using 3-space tabs (a compromise in the developer holy wars
of tabs being 2 vs. 4 spaces), and an IDE that often mixed spaces with
tabs for indentation.  Recent development uses 4-space indentation and
no tabs to maximize formatting consistency.

This commit converts tabs in include, samples (posix, win32 and common)
and src to either 3-space tab stops (to preserve existing formatting)
or 4-space tab stops when possible (as the new preference for
indentation).  This should make it easier to make updates to these
files without needing to change editor settings to match each file's
settings.